pydecenttree.cpp - disabled numpy support because PyArray_Check() was segaulting. Also: 1. Added temporary log hack so I will know when I've    got OpenMP support turned on properly (a bit tricksy on Macintosh,    where it's disabled in the default C compiler).    But how do I force the use of a C compiler that will support OpenMP?!    In setup.py? 2. Now honours "verbosity" argument (-> verbosity variable)    (for now, logs a bit more info when verbosity > 1, but I'll    remove that soon)
This is working for FLAT sequences of double.
Like, say [0,1,1, 1,0,1, 1,1,0].
But it also needs to work for "2D" sequence of sequences.
Like, say [[0,1,1], [1,0,1], [1,1,0]]
And it would be NICE if it worked for numpy arrays too.
But for now calling numpy functions doesn't seem to work?
Is the wrong version of numpy getting linked?! Or what?!
Haven't found a workaround for that issue yet.